+++ internal/migrations/000004_node_engine_inbounds.down.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `nodes`
+  DROP COLUMN `inbounds`,
+  DROP COLUMN `engine`;